First working commit iwth new windings
