+++ femmt/examples/new_winding_example.py
@@ -9,7 +9,7 @@ working_directory = os.path.join(os.path.dirname(__file__), "inductor")
 if not os.path.exists(working_directory):
     os.mkdir(working_directory)
 
-geo = fmc.MagneticComponent(component_type=ComponentType.Inductor, working_directory=working_directory)
+geo = fmc.MagneticComponent(component_type=ComponentType.Transformer, working_directory=working_directory)
 
 core_db = ff.core_database()["PQ 40/40"]
 
@@ -24,19 +24,27 @@ air_gaps.add_air_gap(AirGapLegPosition.CenterLeg, 90, 0.0005)
 geo.set_air_gaps(air_gaps)
 
 isolation = model.Isolation()
-isolation.add_core_isolations(0.001, 0.001, 0.004, 0.001)
-isolation.add_winding_isolations(0.0005)
+isolation.add_core_isolations(0.001, 0.001, 0.001, 0.001)
+isolation.add_winding_isolations(0.0001, 0.0001, 0.0001)
 geo.set_isolation(isolation)
 
 winding_window = model.WindingWindow(core, isolation)
-# top_left, top_right, bot_left, bot_right = winding_window.split_window(fmt.WindingWindowSplit.HorizontalAndVerticalSplit, 0.5, 0.5)
-complete = winding_window.split_window(WindingWindowSplit.NoSplit)
+top_left, top_right, bot_left, bot_right = winding_window.split_window(WindingWindowSplit.HorizontalAndVerticalSplit, 0.5, 0.5)
+#left, right = winding_window.split_window(WindingWindowSplit.HorizontalSplit)
 
 conductor1 = model.Conductor(0, Conductivity.Copper)
-conductor1.set_solid_round_conductor(0.0013, ConductorArrangement.Square)
-complete.set_winding(conductor1, 9, None)
+conductor1.set_solid_round_conductor(0.0005, ConductorArrangement.Square)
+top_left.set_winding(conductor1, 5, WindingScheme.FoilHorizontal, WrapParaType.FixedThickness)
+
+conductor2 = model.Conductor(1, Conductivity.Copper)
+conductor2.set_solid_round_conductor(0.0005, ConductorArrangement.Square)
+bot_left.set_winding(conductor2, 5, None, None)
+
+right = winding_window.combine_vww(top_right, bot_right)
+right.set_interleaved_winding(conductor1, 5, conductor2, 5, InterleavedWindingScheme.HorizontalAlternating)
 
 geo.set_winding_window(winding_window)
 
 geo.create_model(freq=100000, visualize_before=False, save_png=False)
-geo.single_simulation(freq=100000, current=[4.5], show_results=True)
+#geo.single_simulation(freq=100000, current=[4.5], show_results=True)
+geo.single_simulation(freq=100000, current=[4.5, 4.5], phi_deg=[0, 180], show_results=True)
